+++ src/main/java/cat/udl/eps/softarch/mypadel/utils/CourtFilter.java
@@ -1,12 +1,12 @@
-package cat.udl.eps.softarch.mypadel.utlis;
+package cat.udl.eps.softarch.mypadel.utils;
 
 import cat.udl.eps.softarch.mypadel.domain.Court;
 import cat.udl.eps.softarch.mypadel.domain.Reservation;
 
 import java.util.List;
 
-import static cat.udl.eps.softarch.mypadel.utlis.CourtTypeMatcher.matchCourtType;
-import static cat.udl.eps.softarch.mypadel.utlis.ReservationOverlappingChecker.checkDateCollision;
+import static cat.udl.eps.softarch.mypadel.utils.CourtTypeMatcher.matchCourtType;
+import static cat.udl.eps.softarch.mypadel.utils.ReservationOverlappingChecker.checkDateCollision;
 import static java.util.stream.Collectors.toList;
 
 public class CourtFilter {
